Now deserializes STEP files!  All unit tests pass.
It is a heap of spaghetti code though, and requires a lot of refactoring (TODO).

The automatically generated (via xsd) schema is slightly broken, it tends to choose a particular sub-class as the expected type of a property rather than the super abstract class.  A couple of properties have been fixed, there are undoubtedly others out there.