Moved AssertIsMinimumExample method to Utilities.cs
Added the start of an event based rule engine.  Trying to avoid an anaemic domain model by making the rules as methods of the Schema classes, responsible for checking their own properties.  Requires a *lot* more thought & work!

Added an *experimental* JsonSerializer class.  This is based on a modified version of Json.Net, patch attached.